requirements.txt

alembic==1.8.1
click==8.1.3

Flask==2.2.2
Flask-Admin==1.6.0
Flask-Caching==2.0.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2

pandas==1.5.1
psycopg2==2.9.5
psycopg2-binary==2.9.5

python-dotenv==0.21.0

SQLAlchemy==1.4.42
Werkzeug==2.2.2
WTForms==3.0.1
